requirements.txt
fastapi==0.115.11
fastapi_utils==0.8.0
pydantic==2.10.6
uvicorn==0.34.0
pyjwt==2.8.0
passlib==1.7.4
pytz==2025.1
typing_inspect==0.9.0